// rs-x-ui/lib/events/key-code.enum.ts
export enum KeyCode {
   Backspace = 8,
   Tab = 9,
   Enter = 13,
   Shift = 16,
   Ctrl = 17,
   Alt = 18,
   Pause = 19,
   CapsLock = 20,
   Escape = 27,
   PageUp = 33,
   PageDown = 34,
   End = 35,
   Home = 36,
   LeftArrow = 37,
   UpArrow = 38,
   RightArrow = 39,
   DownArrow = 40,
   Insert = 45,
   Delete = 46,
   Zero = 48,
   One = 49,
   Two = 50,
   Three = 51,
   Four = 52,
   Five = 53,
   Six = 54,
   Seven = 55,
   Eight = 56,
   Nine = 57,
   A = 65,
   B = 66,
   C = 67,
   D = 68,
   E = 69,
   F = 70,
   G = 71,
   H = 72,
   I = 73,
   J = 74,
   K = 75,
   L = 76,
   M = 77,
   N = 78,
   O = 79,
   P = 80,
   Q = 81,
   R = 82,
   S = 83,
   T = 84,
   U = 85,
   V = 86,
   W = 87,
   X = 88,
   Y = 89,
   Z = 90,
   LeftWindowKey = 91,
   RightWindowKey = 92,
   SelectKey = 93,
   a = 97,
   b = 98,
   c = 99,
   d = 100,
   e = 101,
   f = 102,
   g = 103,
   h = 104,
   i = 105,
   j = 106,
   k = 107,
   l = 108,
   m = 109,
   n = 110,
   o = 111,
   p = 112,
   q = 113,
   r = 114,
   s = 115,
   t = 116,
   u = 117,
   v = 118,
   w = 119,
   x = 120,
   y = 121,
   z = 122,
   NumLock = 144,
   ScrollLock = 145,
   SemiColon = 186,
   EqualSign = 187,
   Comma = 188,
   Minus = 189,
   Period = 190,
   ForwardSlash = 191,
   BackwardSlash = 220,
   GraveAccent = 192,
   OpenBrace = 219,
   CloseBrace = 221,
   SingleQuote = 222,
}

export enum NumpadKeyCode {
   NumPad0 = 96,
   NumPad1 = 97,
   NumPad2 = 98,
   NumPad3 = 99,
   NumPad4 = 100,
   NumPad5 = 101,
   NumPad6 = 102,
   NumPad7 = 103,
   NumPad8 = 104,
   NumPad9 = 105,
}

export enum FunctionKeyCodes {
   F1 = 112,
   F2 = 113,
   F3 = 114,
   F4 = 115,
   F5 = 116,
   F6 = 117,
   F7 = 118,
   F8 = 119,
   F9 = 120,
   F10 = 121,
   F11 = 122,
   F12 = 123,
}